requirements.txt
fastapi==0.65.3
uvicorn==0.14.0
starlette==0.14.2
pydantic==1.8.2
PyYAML==5.4.1